+++ src/main/java/de/dm/intellij/liferay/language/osgi/config/LiferayConfigSyntaxHighlighterFactory.java
@@ -1,0 +1,17 @@
+package de.dm.intellij.liferay.language.osgi.config;
+
+import com.intellij.openapi.fileTypes.SyntaxHighlighter;
+import com.intellij.openapi.fileTypes.SyntaxHighlighterFactory;
+import com.intellij.openapi.project.Project;
+import com.intellij.openapi.vfs.VirtualFile;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+public class LiferayConfigSyntaxHighlighterFactory extends SyntaxHighlighterFactory {
+
+	@NotNull
+	@Override
+	public SyntaxHighlighter getSyntaxHighlighter(@Nullable Project project, @Nullable VirtualFile virtualFile) {
+		return new LiferayConfigSyntaxHighlighter();
+	}
+}