+++ changelog.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ## 0.0.26
 - Feature: add inspection for  `setUserId()` should not be used in `update` methods in `LocalServiceImpl` classes (avoid unexpected permission expansions)
-
+- Feature: add syntax highlighting, bracket matching and folding for OSGi `.config` files
 
 ## 0.0.25
 - Feature: code completion for the `unbind` attribute in a `@Reference` annotation